Drastic 2.0 update
The ShellMenu class is now instance based rather than static, allowing it to retain state.

Theming options are no longer manually entered for each method, instead a ShellDisplayOptions object is used, this encapsulates all the available theming options.

Fixed bug where cursor would not hide when in DrawMenu mode.

Drastically improved documentation.

Moved to the ApplyLeftOffset function for processing the left offset of displayed text, this is line terminator aware and so the user can display multi-line content and it will be correctly left offset.